# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/remainingnumber/reserveleave/TempReserveLeaveMngRepository.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/workrecord/erroralarm/EmployeeDailyPerErrorRepository.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/repository/remainingnumber/JpaTempReserveLeaveMngRepo.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/workrule/closure/Closure.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.infra/src/main/java/nts/uk/ctx/bs/employee/infra/repository/employee/mngdata/EmployeeDataMngInfoRepositoryImp.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pub/src/main/java/nts/uk/ctx/bs/employee/pub/employee/EmployeeInfoExport.java

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.app/src/main/java/nts/uk/ctx/at/function/app/command/dailyperformanceformat/RemoveAuthorityCommandHandler.java
@@ -48,10 +48,11 @@ public class RemoveAuthorityCommandHandler extends CommandHandler<RemoveAuthorit
 
 		this.authorityFormatMonthlyRepository.remove(companyId,
 				new DailyPerformanceFormatCode(command.getDailyPerformanceFormatCode()));
-
-		this.authorityFormatInitialDisplayRepository.remove(companyId,
-				new DailyPerformanceFormatCode(command.getDailyPerformanceFormatCode()));
 		
+		if(this.authorityFormatInitialDisplayRepository.checkExistDataByCompanyId(companyId)) {
+			this.authorityFormatInitialDisplayRepository.remove(companyId,
+					new DailyPerformanceFormatCode(command.getDailyPerformanceFormatCode()));
+		}
 		this.authorityDailyPerformanceFormatRepository.remove(companyId,
 				new DailyPerformanceFormatCode(command.getDailyPerformanceFormatCode()));
 	}
